BaseLine
Remade previous repository due to public display of Twitter API Keys. My
API keys are being imported by pickle, but you can simply replace the
text on lines 17-20.

This code is based on two tutorials that I found online:
https://www.raspberrypi.org/learning/morse-code/worksheet/
and
http://www.makeuseof.com/tag/how-to-build-a-raspberry-pi-twitter-bot/

Building upon these, by running MorseReciever.py (I know it's spelt
wrong!) with the command line argument of "True", the program allows you
to key in Morse on pin7 to be tweeted.

This Morse is translated letter by letter and printed to the terminal.
The inter-letter gap in seconds is set on line 23, with the inter-word
gap on line 24.

Upon keying in the "End of Transmission" code : "...-.-", the code is
Tweeted to the account details provided on lines 17-20 (providing the
program was run with 'True' as the first command line argument. If not,
then the program runs in "Training Mode" and only clears the output
buffer upon "EoT"

If more codes are created for new letters/transmission elements, you may
add these to the "morse_lookup.py" dictionary file in the manner of the
existing elements.